requirements-dev.txt
pytest>=7.4.0
pytest-cov==4.1.0
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
requests==2.31.0
pytest-mock>=3.11.0
moto>=4.2.0
boto3-stubs[dynamodb,lambda]>=1.34.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
